Change menu location configuration
Instead of specifying the parent slug, let's be more specific about the
actual function that should be called. By default, we stick with the
Appearance menu, as before, with `add_theme_page`. But now, any themes
or plugins that want to locate our dependencies plugins page elsewhere
has to explicitly pass in a config string of `add_submenu_page` or
`add_foo_page()`. This would be outside our library, making the library
itself fine for use in WPORG themes, but the customisation can be caught
by the Theme Check plugin.

See #460, hat-tip @Otto42.